requirements.txt
scapy
rich
PyYAML